Git utility class

+++ app/plugins/module/github.py
@@ -30,53 +30,57 @@ class Provider(BaseProvider('module', 'github')):
 
 
     def initialize_instance(self, instance, created):
-        create_deploy_key = False
-
-        instance.config['username'] = 'git'
-
-        if self.field_remote:
-            instance.config['remote'] = self.field_remote if '/' in self.field_remote \
-                else "{}/{}".format(self.github_org, self.field_remote)
-
-        instance.config['name'] = instance.config['remote'].split('/')[1] if not self.field_name \
-            else self.field_name
-
-        instance.name = instance.config['name']
-        instance.remote = "{}@github.com:{}.git".format(instance.config['username'], instance.config['remote'])
-
-        if not instance.secrets.get('private_key', None) or not instance.secrets.get('public_key', None):
-            private_key, public_key = SSH.create_ecdsa_keypair()
-            instance.secrets['private_key'] = private_key
-            instance.secrets['public_key'] = public_key
-            create_deploy_key = True
-
-        repo, repo_created = self._get_repository(instance)
-
-        if create_deploy_key:
-            if 'deploy_key' in instance.variables:
-                repo.get_key(instance.variables['deploy_key']).delete()
-
-            deploy_key = repo.create_key(
-                "@{}:{}".format(self.github.get_user().login, settings.APP_NAME),
-                public_key,
-                False
-            )
-            instance.variables['deploy_key'] = deploy_key.id
-
-        if repo_created:
-            self._provision_template(instance)
-            self.init()
-            self.commit("Initial template module files")
-            self.push()
-        else:
+        if not settings.GITHUB_TOKEN:
             super().initialize_instance(instance, created)
+        else:
+            create_deploy_key = False
+
+            instance.config['username'] = 'git'
+
+            if self.field_remote:
+                instance.config['remote'] = self.field_remote if '/' in self.field_remote \
+                    else "{}/{}".format(self.github_org, self.field_remote)
+
+            instance.config['name'] = instance.config['remote'].split('/')[1] if not self.field_name \
+                else self.field_name
+
+            instance.name = instance.config['name']
+            instance.remote = "{}@github.com:{}.git".format(instance.config['username'], instance.config['remote'])
+
+            if not instance.secrets.get('private_key', None) or not instance.secrets.get('public_key', None):
+                private_key, public_key = SSH.create_ecdsa_keypair()
+                instance.secrets['private_key'] = private_key
+                instance.secrets['public_key'] = public_key
+                create_deploy_key = True
+
+            repo, repo_created = self._get_repository(instance)
+
+            if create_deploy_key:
+                if 'deploy_key' in instance.variables:
+                    repo.get_key(instance.variables['deploy_key']).delete()
+
+                deploy_key = repo.create_key(
+                    "@{}:{}".format(self.github.get_user().login, settings.APP_NAME),
+                    public_key,
+                    False
+                )
+                instance.variables['deploy_key'] = deploy_key.id
+
+            if repo_created:
+                self._provision_template(instance)
+                self.init()
+                self.commit("Initial template module files")
+                self.push()
+            else:
+                super().initialize_instance(instance, created)
 
 
     def finalize_instance(self, instance):
-        repo, repo_created = self._get_repository(instance, create = False)
+        if settings.GITHUB_TOKEN:
+            repo, repo_created = self._get_repository(instance, create = False)
 
-        if repo and 'deploy_key' in instance.variables:
-            repo.get_key(instance.variables['deploy_key']).delete()
+            if repo and 'deploy_key' in instance.variables:
+                repo.get_key(instance.variables['deploy_key']).delete()
 
         super().finalize_instance(instance)
